MOD: fixed lots and lots of compiler complaints after adding pedantic flags

+++ src/huffman/mod.rs
@@ -60,7 +60,7 @@ pub fn codes_from_lengths_and_symbols(lengths: &[usize], symbols: &[u16]) -> tre
 	for i in 0..lengths.len() {
 		let len = lengths[i];
 		if len > 0 {
-			codes.insert(bit_string_from_code_and_length(next_code[len], len), symbols[i] as u16);
+			codes.insert(bit_string_from_code_and_length(next_code[len], len), symbols[i]);
 			next_code[len] += 1;
 		}
 	}
